+sscanf_test

+++ src/s21_memcmp.c
@@ -8,7 +8,7 @@ int s21_memcmp(const void *str1, const void *str2, s21_size_t n) {
 
   for (s21_size_t i = 0; i < n; i++) {
     if (ptr1[i] != ptr2[i]) {
-      result = ptr1[i] - ptr2[i];
+      result = (int)(ptr1[i] - ptr2[i]);
       break;
     }
   }
