+++ src/s21_strncmp.c
@@ -6,7 +6,7 @@ int s21_strncmp(const char *str1, const char *str2, s21_size_t n) {
 
   for (s21_size_t i = 0; i < n; i++) {
     if (str1[i] != str2[i]) {
-      result = (unsigned char)str1[i] - (unsigned char)str2[i];
+      result = (int)((unsigned char)str1[i] - (unsigned char)str2[i]);
       to_exit = 1;
     }
